Changed how FitParameters reads/sets the binning.
Changed how FitParameters both reads and sets the binning. The function
now takes a string and now can simply construct a new FitBin to add to
the vector.

CCQEBin has been changed to FitBin and given a couple constructors to
allow for vector.emplace_back() and easier assigning.

Modified CMakeLists.txt to only build ingridFit.cc so that as breaking
changes are made, the other code is not compiled.