Further edits to AnySample; New GetSampleBreakdown
More edits to the structure of AnySample. Wrote a new version of
GetSampleBreakdown which now includes the topologies as a function
parameter.

Topologies can be passed as a vector of strings and the function will
then construct properly sized vectors of histograms for each topology.
All the histograms are now constructed as instances rather than a bunch
of pointers with new.

Both XsecFitter and ingridFit had each use of GetSampleBreakdown changed
to the new version. Old version is still in the class for (temporary)
    backwards compatibility.