Cleaned up RNG and Cov Matrix code in fitparam.
Rewrote the functionality to have random prior values for the fit
parameters. The RNDM function in ROOT is old, and shouldn't be used.
Changed to Uniform call and the RNG is no longer a pointer declared
object.

Also had to fix the prefitParam histogram for the fit output. Originally
it just filled all parameters with a 1.0, regardless of the actual
value. Histogram now fills with correct value.

Cleaned up the SetCovarianceMatrix code in AnaFitParameter. Changed
function paramter pass to const ref which achieves the exact same thing.