Minor restructuring. New isSignal() function.
More minor restructuring and rewriting of various parts of the code.

The AnaEvent class now has an m_signal variable, which marks the event
as a signal event or not. The AnyTreeMC class now takes a list of signal
topologies as part of reading the events, and marks each event as signal
or not.

This now can be used by other classes, such as the FitParameters, to
determine if a given event is signal rather than hard coding checks to
the event topology.