New functionality to fitparameters!
Many breaking changes have been made to change how the various
parameters in the fit are added and treated.

All samples now have a detector variable which specify which fit
parameters apply to the sample.

Adding fit parameters now require adding detectors to the fit object
which takes the detector specific binning and offset in the covariance
matrix.

A new function InitParameters now builds the parameter list from the
added detector binning and configures the fitter.

When reweighting an event, it now needs to now which detector specific
parameters to use, and will look up the correct binning based on the
offset.

This appears to be kind of working. Extensive tests needed.