First part of new JSON configure parsing
This is the first part of using JSON files to configure the fitter. An
OptParser class has been added to handle the argument parsing for both
JSON, and later command line options to override the JSON values.

Currently only the JSON parsing has been implemented. It is based on
JSON for Modern C++ by Niels Lohmann, which is provided as a header
library.

Inside the main function itself, the JSON values are used to set
everything but the samples themselves. This commit is to save the
working changes before writing a temporary solution to set up the
samples from the JSON file.