Second part of JSON parsing, now functional
The rest of the JSON parsing code has been implemented. Now the samples
are configured from the JSON file including fit/flux parameters.

Small change to how AnySample is designed. Changed flag m_empty to
m_use_sample to accurately reflect the config file and how the flag is
supposed to be used. Instead of setting m_empty to false to fit a
sample, m_use_sample is set to true.

Updates to INGRID binning and commiting the JSON config to have a clean
working directory.