AnySample: Changed where m_hsig is declared/filled.
In AnySample, changed where m_hsig is declared and filled to better
reflect how its used and to allow optional use of GetSampleBreakdown.

GetSampleBreakdown was changed so that it outputs the sample statistics
every time the function is called and no longer is conditional on the
save variable.

A number of minor changes to better practices and style to AnySample, such as
removing histograms with `delete` instead of the ROOT method.

In unrelated changes, a number of files have been deleted to clean up
the directory. CMakeLists has been changed accordingly.