Updates to ToyThrower; Start calcXsec
A few small changes to the ToyThrower class. Updated the error message
when the toy vector is not the same size as the covariance matrix. Also
changed how the vector was set to zero when using std::vector as clear()
removes all elements.

A parameter to specify the output file of the calcXsec code was added to
OptParser.

The initial code for calculating cross sections and propagating error
has been added. It currently does the same setup as the fitter to be
able to reweight events, and tests the toy throwing.

The toy throwing looks to be working. The thrown parameter is the
nominal parameter plus the throw value. It seems to take about 1E6
throws or so for the covariance matrices to converge pretty well.