Profiled calcXsec and faster code
Profiled calcXsec using the chrono library to obtain computation times
for various sections of the code. By far the longest time is reading
events from ROOT trees.

The original throw loop was reading in the data events every time,
taking awhile to do so. A new mode was added to FillEventHisto to avoid
filling the data histogram when not needed, thus avoiding ROOT I/O.

This gave about a factor of 20 speedup in time per throw, and to overall
execution time. The next longest thing is now calculating the
correlation/covariance matrices when using a large number of throws.

Preprocessor define macros have been added to allow switching on/off
debug messages. The makeflags have been added to CMakeLists.