Started adding other xsec calculations.
Started writing code to perform the other throws/normalization
calculations for measuring the cross section.

XsecExtractor is a class to facilitate the other throws and
normalziations needed to calculate the cross section. The design idea is
to initialize one XsecExtractor object for each detector in the
analysis.

Each XsecExtractor object will be responsible for performing the number
of targets, flux integral, bin widths, etc. and will take an event rate
histogram and calculate the cross section.

New OptParser configure options have been added to support the above
class, such as number of targets.