Implemented flux and bin width normalization.
The flux and bin width normalizations were implemented in
XsecExtractor. The bin widths are calculated from a given binning file.

The flux normalization has several methods: giving a flux file for
throwing, giving a flux file and error, or just specifying the integral
and error.

Both the number of targets and flux error now have a scoped enum flag to
denote if the error is given as a relative or absolute error.

The OptParser has been modified to accept the arguments needed to
specify the flux integral and error. The new arguments have been added
to the JSON config file. The binning files have been updated to test how
the bin width affects the extracted cross section.

The code has been added to calcXsec and tested that it runs.