Fixed how multiple samples are handled in CalcXsec
The bulk of this commit is about fixing how detectors with multiple
samples are handled in the calcXsec, error propagation stage.

The signal events for each sample are now filled in a new histogram
using the final analysis binning, and then that histogram has the bin
widths, number of targets, etc. normalizations applied.

The postfit histograms for each detector are then concatenated together
for calculating the covariance. This process has only been written for
the best fit values, not the toy throwing.

New binning files have been added which match certain tech notes.

A number of minor changes to code structure, such as haivng a header
file to reuse helpful structs/classes throughout the code.