CalcXsec now throws xsec normalizations.
When making a toy throw the bin widths, number of targets, and flux
integral normalizations are now applied. The flux and number of targets
are thrown according to their error from the configure file.

Finally fixed a variable uninitialized warning in XsecFitter.cc and a
spurious semi-colon in FitStructs.hh

The code now compiles with zero warnings (or errors) using -Wall and
-Wpedandic.