Changed how binning was set for AnaSample.
Changed how the binning was set for AnaSample to make it uniform with
the rest of the classes. Binning is now set by passing the file name
containing the bins in the constructor.

Also removed all other binnings other than the linear bin index. These
other binnings may be added back in, but only when necessary and will be
implemented in a different way.

AnaSample now uses the FitBin struct for storing bin edges. The
FitStructs header that contains the struct has been moved to inlcude/
and all CMakeLists have been updated to reflect this change.

CalcXsec and ingridFit have been updated to use the new AnaSample
interface.