Added colored text output.
Working on the important things: adding colored text output. The
ColorOutput header contains string constants and enums under the color:: namespace to add color and styling to the output text.

The const strings can be used directly or the ColorText can add color to
any object that can be processed by a stringstream.

Color output has been added to ingridFit so far.

Also finally commiting the clang-format file with a close version of my
style preference.