Implementing XsecParameters Part 1
The AddDetector function to load dials has been added to XsecParameters.
It reads in a dials JSON config and constructs an XsecDial for each and
initializes the values.

The OptParser has been modified to handle the extra inputs for the xsec
dials. XsecParameters now loads json.hpp to perform the dials JSON
config reading.

Better error handling has been added to BinManager and XsecDial. Both
complain if their respective files can not be opened.