Implemented InitParameters for XsecParameters.
Implemented the InitParameters function for XsecParameters. This
differs from the Fit/Flux implementation slightly by using a vector of
the detector strings to control the order in which parameters are added,
    rather than the custom set comparator.

The detector parameter offset is still necessary, although the class now
calculates and keeps track itself, rather than it being a config
parameter.

Next up is to modify InitEventMap to handle storing indices for dials.