InitEventMap and ReWeight for XsecParameters Implemented
The last parts of implementing the XsecParameters have been finished.
The xsec parameters had to be treated a little differently to get the
EventMap and ReWeighting to work compared to the Fit/Flux parameters.

The fit runs and more or less seems okay. The results are a little
funny. For an asimov test for Fit/Xsec parameters, the fit complains
about the second derivative for INGRID_MARES, but the fitted values are
all still nominal.

However when running with a weighted data sample (10% higher than
        nominal), the fitter returns multiple parameter values higher
than 1.1, which is wrong (with or without the xsec parameters). Time to
checkout earlier code and test what went wrong.