Improvements to config and xsec covariance inputs.
A number of new features have been added to the input configuration for
the fitter.

The dials can now be included or skipped by setting the "use_dial" flag
in the dials.json file.

Each detector now has a "use_detector" option to allow for single
detector fits.

The cross section covariance is now a text file input, which is
specified in the config file. This might change to a ROOT file input.

The superTreeConvert macro has been edited to set the reaction variable
for INGRID events based on their NEUT interaction code.

Fixed the precision for standard out being set to only three digits for
all stdio output.