Clang Format DetParameters; Rainbow/Color Text.
Ran the DetParameters files through clang-format to make them easier to
look at. They need to be modified to match the rest of the framework.

Added text color to AnaSample. Wrote a Rainbow Text function to
automagically rainbow-fy anything that can be passed to a string stream.

Changed the random priors for the FitParameters to be a Gaussian number
rather than uniform.