Changed which loop was multithreaded
When reweighting the events during the fit, the loop over events is now
the multithreaded loop instead of the loop over samples.

The event loop shows greater speed up and efficiency over parallizing
the sample loop.

Also, for now the check of the determinant being non-zero for the
covariance matrix has been commented out. It still has issues with the
numerical precision with the determinant from the Invert function being
rounded to zero.