New code to calculate detector covariance.
A new piece of code was written to calculate the detector covariance
from a HighLAND output file. The code also will plot the distribution of
toys for validation.

Yes, the extension is .cxx instead of .cc like the rest of the
repository. Thinking about changing the convention for naming files.

The CMakeLists.txt in src/ has been edited to compile the new code.