Fixed bug when calculating covariance matrix.
Fixed a bug when calculating the covariance matrix which would cause an
incorrect matrix at best, and crash and burn at worse.

When checking if a sample was in the use_samples[] array, the loop would
run off the end of the array either causing undefined math to happen, or
a segfault.

The method for checking if a sample is in the use_samples[] array is now
performed by std::count() checking if a sample is in the array. On a
related note, the problem would have been found quicker if the .at()
    method was used for access.

Unrelated, the histograms axes now are scaled so that the entire
distribution of toys fits within the axes.