Many misc. edits/revisions. Removed all regularization code.
Finally removed the rest of the code to handle regularization. This will
be added back in once it becomes necessary. The same thing will happen
to the old toy-throwing code eventually.

Removed the FSI and NuclFSI parameter classes/files. They were redundent
with the XsecParameters class. Updated CMakeLists.txt to remove the
dependencies on those files.

Renamed AnyTreeMC to AnaTreeMC to match the rest of the anaevents
classes. Updated CMakeLists.txt with the new name.

Originally when storing the True and Reco Enu values they were divided
by 1000 to convert from MeV to GeV. This has been removed so all
variables are in MeV in the AnaEvent class. Classes/functions that need
variables in GeV are now responsible for the conversion.

For example, the Flux parameters now do the conversion from MeV to GeV
when passing the value to GetBinIndex().

Removed a couple files from the macros directory; will eventually delete
more.

Edited ingridFit and calcXsec to match the changes to
anaevents/fitparam/xsecfit.

Probably a few more misc. changes that I have missed.