Fixed bug in AnaTreeMC; Renamed some files.
Fixed a bug which caused compilation to fail in AnaTreeMC. Shame on me
for pushing a commit which did not compile.

Renamed applications to fit naming scheme. All executables are now prefixed
with "xsllh". CMakeLists.txt has been updated with the name changes and
the instructions to compile ccqefit, fitbase, and propError have been
removed since they are not needed.