Added ProgressBar, Updated ColorOutput.
Fixed the RainbowText function not rainbow-fying the entire text
properly. There was a bug when calculating when to switch to a new color
where it would always switch after RAINBOW_SIZE numebr of characters,
which could eventually lead to walking off the array.

Added a ProgressBar class. It can be configured with variable length,
character for the bar, prefix text, and rainbow colors. A single
Print function then prints the formated text to the terminal.