New Tree Converter code/executable.
The superTreeConvert.cc macro has been replaced by the xsllhTreeConvert
code. The new code is compiled code which uses a JSON configure file to
specify the ND2/ND5 trees to convert. More options will probably be
added to the configure file.

The new code gives almost identical results as the original
superTreeConvert macro, which is considered good enough to make the
switch.

Changes still need to be made to the fitter, specifically AnaTreeMC, to
adapt the small changes to the output tree (mostly variable names). In
addition the output tree currently does not have the muon/proton kinematics as
extra variables.

Small update to ProgressBar: it now prints a newline character when the
bar reaches 100%.

CMakeLists.txt updated to include the compilation of xsllhTreeConvert.
Also included is xsllhUnitTests since CMake will try to compile it. It
currently does nothing except print some text.