Fixed critical bug in AnaSample. Removed ToyThrowing code from AnaFit.
Found and fixed a critical bug in AnaSample. The TTree variable names
had not been updated when reading in the events from the data tree. This
caused all the events to be assigned to sample 0, bin 0. This has been
fixed.

Removed unused code from AnaFitParameters and removed
FluxParameters_norm. The biggest removal was all the ToyThrowing code
and GeneratingToyData code. These will be added back in when needed,
using the newer routines for ToyThrowing developed for the error
propagation.

A number of smaller changes were made to AnaFitParameters to update the
style, better encapsulate the code, and remove more unused code.

All derived fit parameters are entirely responsible for the variables
and code to AddDetectors and initialize their parameters. AnaFit only
holds the parameter vectors and covariance matricies.

XsecParameters was updated to remove the corresponding code from
AnaFitParameters.