Adding eigen decomposition code backend.
The backend code that performs the eigen decomposition has been written
and tested. It appears to be working correctly.

It is based on the BANFF implementation and can optionally use the eigen
decomposition or singular value decomposition (SVD) of the covariance
matrix. The code currently defaults to using the eigen decomposition
method.

It has been templated to accept both TMatrix and TMatrixSym as an input.

CMakeLists.txt has been updated to compile the code.