Updating EigenDecomp with better methods to undecompose.
EigenDecomp has new routines to undecompose only a part of a parameter
vector or matrix.

These build a temporary set of eigen vectors to only transform the
selected part. These require the start index of the parameters in the
vector or matrix

Splitting this from the next commit to better split the types of
changes.