Implemented eigen decomposition inside fit parameters.
The ability to perform eigen decomposition has been implemented in the
fit parameter classes.

AnaFitParameters holds nearly all the code to handle the decomposition.
The only thing the individual parameter classes are currently
responsible is initializing the parameters when decomposed.

Inside XsecFitter.cc the parameters are transformed back to the original
space each iteration if needed. One the fit is finished, the parameters
and final covariance matrix are transformed back and saved.

Tiny update to ProgressBar, adding function to set max progress.

Tiny update to the master configuration file. Added configuration
file for the new tree converter.