Added decomp flags to OptParser, new functions in EigenDecomp.
Added in flags for eigen decomposing parameters to OptParser and the
master configuration file. Each parameter set can be individually set to
be decomposed or not.

Added a function to calculate the number of eigen values needed to
retain some specified fraction of information/variance.

Modified xsllhDetVariations to set diagonal covariance entries of zero
(or less) to 100% error to allow the matrix to be invertible. Also now
uses the progress bar for output.