New capability to automatically fix vars w/ small eigenvalues.
Added the capability fix parameters with small eigenvalues in the fit.

AnaFitParameters now has a pars_fixed vector of boolean values. When the
fit is initialized, it will check each parameter if it needs to be
fixed.

An information fraction variable has been added to AnaFitParameters
which sets the cut for keeping eigenvalues. The parameter sets then
check how many parameters they need and set the params_fixed vector
accordingly.

XsecFitter now uses the params_fixed vector to fix the corresponding
variables just after initialzing the prior value.

The configuration file has been modified to include the cut values for
keeping parameters, and OptParser now reads the values.

Small change to EigenDecomp::GetInfoFrac(), it now always reports the
number of parameters needed for a given fraction. So if a bad value is
given, it will report zero parameters or all of them (depending on exact
bad value).