Cleaned up XsecFitter to streamline code.
Numerous small changes to XsecFitter to streamline the code and to
better match the style of the rest of the code.

Removed a number of member variables which were not needed. Renamed most
of the output histograms.

Changed the constructor to require the pointer to the TDirectory for
output. The SetSaveMode function has been changed to SetSaveFreq which
now only controlls how often the intermediate histograms/parameters are
saved.