Added back regularisation to the fit.
Implemented regularisation back into the fit. Each parameter set is now
responsible for its regularisation, if any. The CalcLikelihood function
in XsecFitter asks for the regularisation penalty from each fit
parameter set for which regularisation is enabled.

XsecFitter similarly stores the chi2 per iteration for the
regularisation and saves the information as a histogram to the final
output.

AnaFitParameters now has several variables to control/enable the
regularisation. It contains a virtual function for CalcRegularisation
which returns zero in case the function is called for a derived class
that has not defined it.

FitParameters currently implements two forms of regularisation: L1 and
L2 regularisation.

OptParser has been modified to include parameters for the
regularisation.