Reimplemented toy throwing capability.
The bulk of this commit is reimplementing the ability to make toy throws
and fit the toy data.

The old ThrowParms code has been deleted, and the ToyThrower code is
used for making the throws.

AnaFitParameters contains the function to make a toy throw based on its
original covariance and original prior parameters. If the parameter set
has no covariance (e.g. the template parameters), it currently just
returns the original priors.

XsecFitter now has a function to generate toy data from the fit
parameters it knows about. It generates a toy set of parameters and then
stores the reweighted distribution in the data histgoram. The prediction
histogram is reset to the original prior values. Statistical
fluctuations can be applied to the toy throw.

The toy throwing code works even for decomposed parameters. The toy
throws are always based off the original covariance and priors. The
parameter sets now always save the original priors seperately for this
reason.

New function in XsecFitter to remove the systematic error term to the
chi2 in the likelihood. This is to allow the systematic parameters full
freedom to move and fit the data when testing the fit machinery.