More config arguments. New enum. More saved histograms.
Added a number of new config arguments. New flags to turn on/off stat
fluctuations or systematic error from the fit parameters. Also a
variable to define what type of fit to be run.

OptParser has been updated to handle the new options. Work in progress
to translate strings to enums.

New enum values for the fit type.

Histograms of the original, undecomposed, prior parameters are now saved
in the same format as the final parameters for easy comparisons.