New config option: do_throw
Added another config option for running the fit. When running a toy
throw of the parameters, each parameter set can chosen seperately to be
thrown or not. This allows for fits where only the flux parameters where
thrown, for example.

New scripts/ directory to store various scripts for running the fit.
Currently contains a script to submit jobs to the cluster. Did not want
to use the macros directory.

Random change from DEBUG_MSG to DEBUG trying to get some CMAKE stuff to
work.

Deleted some more files because I felt like it.