Slight changes to XsecDial and spline configuration
Changes to XsecDial to prepare for using ptheta splines in addition to
q2 splines.

The GetSplineIndex function now has an overload where it takes a vector
of indicies and a vector of kinematic variables. The BinManager still
handles the kinematic bin index the same way, but the other indices are
multipled by their specified dimension to calculate the final spline
index.

The SetDimensions function now has an overload where it takes a vector
of integers to specify the widths of each bin/label/etc. Using this
method requires the dimensions to be precalculated.

The indices for GetSplineIndex must be in the same order as passed to
SetDimensions.

The dials.json file has been renamed to q2_dials.json, and now there are
seperate directories for the different splines.