New Xsec Covariance builder.
New executable to build a ROOT TMatrix covariance from a text file for
the cross section parameters.

It reads in a text file of the covariance and stores it in a TMatrixTSym
by default. There are a number of command line options to perform
additional operations.

The code can calculate the correlation matrix, add offset terms to the
diagonal, check if the matrix is invertible, build the INGRID
covariance, and mask rows to remove parameters.

The code is still a bit rough, and the option to store as a TMatrixT
instead of a TMatrixTSym has not been implemented yet.