Testing with ptheta splines.
Committing the changes to ensure a working directory before modifying
the code futher.

The code is currently set up to read/use the ptheta splines. Updating
the config.json file and adding the ptheta_dials.json file.

Removed a few files that were unnecessary.

Slight change to CMakeLists file to set the version of the code to
0.9.0.