Changed xsec covariance loading from text to ROOT file.
Changed the cross section covariance loading from text file to ROOT file
(using TMatrixTSym). The fit now expects a ROOT file in the
configuration file.

This was done to match the flux and detector covariance matrices. The
xsllhXsecCov executable can be used to produce a ROOT file from a text
file.