CMake checks compiler version. New options for configure.
Futher updates to the CMake build. The project now requires CMake 3.5 or
higher.

CMake checks for a minimum version of the compiler, needing 4.8.5 or
higher for GCC and 3.2 or higher for Clang.

A couple new options were added for configuring the build: an option to
enable the -march=native flag and enable/disable -Wall.

Also included in this commit is a tiny update to submit_fit_job.sh to
run the job file from the scripts/ directory.