Massive cosmetic update. Updated README.
Remade the README file to be relevant, and slightly useful. It at least
contains instructions on downloading with git and how to build. It
provides a base to add further instructions/information.

Otherwise many cosmetic changes in many files. Most files had their TAG
variable defined and colored. Now most of the code has proper colored
output; color scheme subject to change.

Slight fix when reporting the number of eigenvalues kept when performing
PCA. It was off by one.

New preprocessor define to disable colored output if desired. This will
be added as a CMake option once the updated cmake branch is merged.

OptParser now has a PrintOptions() function to display some of the configured
options. This is now used in xsllhFit to display the options at the
start of running.