Added functionality to combine samples in DetVar code.
Added the functionality to combine Highland branches in to a sample when
calculating the detector covariance.

This works by specifying the mapping between samples and branches
through a JSON object, which then gets translated in to a C++ map.

The keys are the sample numbers for the analysis, and the values are
vectors of ints containing the cut branches to use for the sample.

To have a sample with branches 0,1 and another sample with branches 2,3
it would look like this in the JSON:
"samples" : {"0": [0,1], "1": [2,3]}

Note that JSON object keys must be strings.

Also added flag to enable/disable using files when running the detector
covariance.