New functionality for DetVar: separate sample binning.
Added yet more functionality for the detector covariance, and fixed a
major bug in the process.

Each sample now has its own binning file. This is specified as a JSON
map where the keys are the sample number and the value is the path to a
binning file. All branches in a given sample use the same binning.

The code now properly calculates the number of toys to use when building
the covariance matrix. Each file can have a different number of toys,
but to create the covariance each sample needs to have a common
number. Therefore the lowest number of toys common to each file are
used when creating the matrix.

Also added the ability to input the name of the covariance/correlation
matrix.

The input JSON file has been updated.