Rewrote TreeConvert.
Rewrote the tree converter code to have similar capabilities to the
detector covariance code.

The tree converter now can run over multiple highland files. Each file
specifies its cuts, branches, and how to combine the branches in to
samples. The sample label can be any integer.

The minimum necessary variables are also specified in the configure
file. More variables or cut conditions can be added in the code if
desired, but the variables in the configure file must be specified for
the fitter to run.

Analysis specific details will be left to the analyzer to implement.