Added target variable, added dimension config option.
Added the target variable to AnaEvent and AnaTreeMC. Inclues SetTarget()
and GetTarget() functions.

Finally renamed RecD1 and RecD2 to RecoD1 and RecoD2 everywhere.

Added the dimension config option to the spline configure file. This
sets the widths for the array when indexing/reading the splines. It is
specified as an array, e.g. [580, 58].

A global dimension can be set for every dial, and each dial can have its
own dimension array which will override the global option.