Adding functionality for multiple signal types.
Added the ability to define multiple signals per detector, i.e. CC0pi
Carbon and CC0pi Oxygen for ND280.

The FitParameters now takes a vector of signal defintions and keeps
track of each signal type in the same way the detector parameters keep
track of each sample.

Each signal type has its own defition and binning. Signal types are
defined under each detector in the configure file.

The AnaTree reader is still responsible for tagging events as signal,
so it also reads the vector of signal definitions. It then labels each
event as either signal or not, and assigns it an integer signal type (or ID)
based on which definition it passed. The ID is simply based on the order
the signals were defined in the configure file.

Found a bug when compiling in DEBUG mode where some asserts fail in the
JSON parser when trying to read the dimensions for the splines. Removed
the try/catch block for now.

Noticed a (potential) bug when calculating the regularization for the fit
parameters. The regularization runs over all the fit parameters, which
would cause it to regularize adjacent bins which belong to different
signal definitions. Not sure if this is correct behavior.

The error propagation code is now definitely broken with the changes to
the fit parameters. Need to rewrite entirely. It's a mess.