Removing old configure options. Deprecating CalcXsec.
Removed the signal topology option from the configure file and
OptParser. It has been replaced by the new signal definition method.

Deprecating the current version of the error propagation code for
calculating the cross section. It won't function as is with the changes
to the signal definition. Needs to be rewritten.

Also adding preprocessor defines to detect if OPENMP was successfully
detected for including omp.h to prevent compilation errors.