Added successful matrix inversion check.
AnaFitParameters now checks for successful inversion of the covariance
matrix by inverting with TDecompLU::InvertLU and checking the status
flag.

This is more accurate than simply checking if the determinant is zero.
Also, when calling Invert through TMatrixT it will not report if the
inversion was successful.

If the inversion is successful, then the inverted matrix is copied to
covarianceI variable from the temp matrix. A temp TMatrixD is needed
because InvertLU specifies a TMatrixD in the function call.