New ErrorProp code started. Only does initialization.
The new error propagation code has been started. The original
xsllhCalcXsec has been completely gutted to serve as a test bench for
the new code.

The majority of the code related to running/managing the error
propagation is in errprop/ and is compiled as another library.

It currently consists of two classes: FitObj and XsecCalc (both names
subject to change).

FitObj is responsible for setting up the fit machinery and performing
reweighting of events. It takes the exact same config file as the fit,
and does everything except run the fit (and load covariance matrices).

XsecCalc manages the toy throwing and will eventually take care of the
normalizations, e.g. number of targets.