ErrProp code successfully fills signal histograms.
The error propagation code is able to initialize and fill the signal
histograms for both the selected and truth trees. Also works with
multiple signal definitions.

Sight change to FitParameters to remove the static variable from
AddDetector; the signal id is now a class variable. Probably should have
been that way from the beginning.

Slight change to xsllhFit to not print any sample information if the
sample is a truth sample. The fit should never use the truth tree and it
is confusing to see the info get printed if it's not used.