Yet another config file (for error propagation).
The error propgataion, or more specifically XsecCalc, has its own config
file. This will cause less confusion and be easier to maintain rather
than trying to have one config file for both the fit and the error prop.

Eventually the unnecessary options from the fit config will be removed.

XsecCalc now reads a JSON config file and initializes the fit objects
using the specified files in its own config.

XsecCalc reads in the normalization parameters (flux and number of
targets) for each signal present in the fit config file. If a
signal is present in the fit config, but not the error config, then an
exception is thrown and caught, causing the program to exit.

If the other way around occurs, then the signal norm is simply ignored.

A struct was made to carry the normalization parameters for each signal,
stored in a vector with the ID as the index.