ErrProp code now reads fit file. Saves output.
XsecCalc now reads the postfit parameters and covariance matrix from the
fit file. It always reads the one in the config file when constructed,
but a new file can be read with ReadFitFile(). xsllhCalcXsec has
commandline options to override the input/output file choices in the
JSON config.

XsecCalc now has a function to save its output, which right now just
saves the postfit variables and a couple signal histograms.

FitObj now has a function to reset the signal histograms before filling.
This is necessary since the signal histograms persist after the function
call, rather than being created and returned within the function.