ErrProp event reweighting. ToyThrower initialized.
XsecCalc can use the FitObj to reweight events according to some
parameter vector. Specific functions exist to reweight to either the
nominal or best fit parameters.

The ToyThrower class is now initialized in XsecCalc. ToyThrower now has
the capability to force the covariance matrix to be positive definite.
This repeatedly adds some value passed as a parameter to the diagonal of
the covariance matrix and checks if the decomposition is successful. To
prevent an infinite or near-infinite loop, a iteration limit is imposed.

Slight addition to DetParameters to facilitate the ability to ignore
true events. This is done by having a mode (mode=2) where all events in
the samples are ignored by the ReWeight() function. FitObj uses this
when loading true events. This way the reweight functions and parameter
vectors do not need to change.

ToyThrower has been updated to have TAG/ERR strings and colored text.