ErrProp simple toy-throwing implemented.
XsecCalc has a basic version of the toy throwing code implemented. Toys
are generated with ToyThrower and passed to FitObj for event
reweighting. Currently FitObj concatenates the signal histograms and
returns them to XsecCalc.

The selected and true event histograms are divided to produce an
efficiency.

The best fit versions and all toys are saved for now to test
functionality. The toy throwing will probably change to getting a vector
of signal histograms, appling all normalizations, then concatenating.