ErrProp now calculates cross sections!
XsecCalc now has functions to apply all the normalizations to calculate
a cross section.

Each one takes a single histogram and applys the specified
normalization, making a random throw if requested.

The flux normalization can only apply the flux integral in the config
file for now. Using the flux histogram and reweighting it will be
harder, and a task for later.

The bin width normalization has a parameter to change the units from MeV
to something else when applying the bin width.

The bin widths are calculated using new functions in BinManager, which
return the bin width for a given bin, or given bin for a given
dimension. FitObj now has a function to return a reference to a
BinManager object.

The toy throwing in XsecCalc now gets the vector of signal histograms
from FitObj, applys each normalization separately for each signal, and
then finally concatenates them together.