ErrProp covariance code added.
XsecCalc now can calculate the covariance and correlation matrix from
the toy throws. It can use either the mean of the toys or the best fit
when calculating the matrix.

Few extra things added. xsllhCalcXsec has the ability to override the
number of toys using the command line. The toy throws are not
automatically saved.

Overall the output does not look right. The post fit correlation matrix
is not nearly as uniformly correlated as expected, and the best fit
cross section is about 2x higher than it should be.

Stay tuned for further testing.