ErrProp, OptParser tiny but critical bug fixed
A tiny but critical bug was found in OptParser in which the variables
for POT were declared as int's, when they should be doubles. This could
cause incorrect values to be loaded from the JSON, and more importantly
leads to calculating the ratio incorrectly.

The data/MC POT ratio is necessary for getting the correct normalization
for both the fit and the cross section calculation. This has been fixed.

In addition, the data/mc POT ratio is now applied to the histograms in
FitObj to get the correct normalization when using differing POT.

The toy throwing code has been verified to work correctly. The issue
with the calculated correlation matrix for the cross sectino is believed
to be due to incorrect parameters when throwing the flux and target
normalizations.

The error propagation config lists the flux and target as the relative
(percent) error, but was being used as if it was the absolute error.
This results in effectively no variation between throws. This has been
fixed by adding another parameter to specify what type of error is being
used.